ws-ldap-check-credentials/src/requirements.txt
flask
dynaconf
pydantic
python-ldap
loguru
gunicorn
